Simplify workflow
Remove the failure detection and re-run logic, as it was
overcomplicated and redundant.
Restructure the workflow to tag the dependecies repos in
the same Job as the remote workflow dispatch.
This ensures that if a remote workflow fails, we can apply
any fixes in the remote repo, trigger a re-run and be certain
that the remote repo tag will be moved to the latest commit.

Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>